synthetic/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pypi_0
async-timeout=4.0.2=pypi_0
attrs=21.4.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
bleach=5.0.0=pypi_0
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.3.29=h06a4308_1
cachetools=5.0.0=pypi_0
certifi=2021.10.8=py39h06a4308_2
cffi=1.15.0=py39hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
cryptography=36.0.0=py39h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
debugpy=1.6.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.4=pypi_0
executing=0.8.3=pypi_0
fancycompleter=0.9.1=pypi_0
fastjsonschema=2.15.3=pypi_0
ffmpeg=4.3=hf484d3e_0
fonttools=4.33.3=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.3.0=pypi_0
fsspec=2022.3.0=pypi_0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.6.6=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.46.0=pypi_0
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipdb=0.13.9=pypi_0
ipykernel=6.13.0=pypi_0
ipython=8.2.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.7.0=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.1.0=pypi_0
jpeg=9e=h7f8727e_0
jsonschema=4.4.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.3.0=pypi_0
jupyter-console=6.4.3=pypi_0
jupyter-core=4.10.0=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=1.1.0=pypi_0
kiwisolver=1.4.2=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
littleutils=0.2.2=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.7=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
multidict=6.0.2=pypi_0
nbclient=0.6.0=pypi_0
nbconvert=6.5.0=pypi_0
nbformat=5.3.0=pypi_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.5=pypi_0
nettle=3.7.3=hbbd107a_1
nflows=0.14=pypi_0
notebook=6.4.11=pypi_0
numpy=1.21.5=py39he7a7128_1
numpy-base=1.21.5=py39hf524024_1
oauthlib=3.2.0=pypi_0
ogb=1.3.3=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1n=h7f8727e_0
outdated=0.2.1=pypi_0
packaging=21.3=pypi_0
pandas=1.4.2=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathtools=0.1.2=pypi_0
pdbpp=0.10.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.1=py39h22f2fdc_0
pip=21.2.4=py39h06a4308_0
prettytable=3.2.0=pypi_0
prometheus-client=0.14.1=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.29=pypi_0
protobuf=3.20.1=pypi_0
psutil=5.9.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydeprecate=0.3.2=pypi_0
pygments=2.12.0=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.8=pypi_0
pyrepl=0.9.0=pypi_0
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.12=h12debd9_0
python-dateutil=2.8.2=pypi_0
pytorch=1.11.0=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-lightning=1.6.3=pypi_0
pytorch-mutex=1.0=cuda
pytz=2022.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=pypi_0
qtconsole=5.3.0=pypi_0
qtpy=2.0.1=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
seaborn=0.11.2=pypi_0
send2trash=1.8.0=pypi_0
sentry-sdk=1.5.10=pypi_0
setproctitle=1.2.3=pypi_0
setuptools=61.2.0=py39h06a4308_0
shortuuid=1.0.8=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.38.2=hc218d9a_0
stack-data=0.2.0=pypi_0
tensorboard=2.9.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
terminado=0.13.3=pypi_0
threadpoolctl=3.1.0=pypi_0
timm=0.5.4=pypi_0
tinycss2=1.1.1=pypi_0
tk=8.6.11=h1ccaba5_0
toml=0.10.2=pypi_0
torchaudio=0.11.0=py39_cu113
torchmetrics=0.8.2=pypi_0
torchvision=0.12.0=py39_cu113
tornado=6.1=pypi_0
tqdm=4.64.0=pypi_0
traitlets=5.1.1=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py39h06a4308_0
wandb=0.12.15=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.6.0=pypi_0
wilds=2.0.0=pypi_0
wmctrl=0.4=pypi_0
xz=5.2.5=h7b6447c_0
yarl=1.7.2=pypi_0
zipp=3.8.0=pypi_0
zlib=1.2.12=h7f8727e_2
zstd=1.4.9=haebb681_0
